backend/requirements.txt
fastapi
uvicorn         # ASGI server to run FastAPI
pymongo         # MongoDB client
python-dotenv   # Load environment variables from .env
